+++ unn/fuzzy_logic/Fuzzy_logic/ModelTSK0.cpp
@@ -13,7 +13,7 @@ ModelTSK0::ModelTSK0(Data _data): train_data(_data)
     std::cout << "Num rules: " << num_rules << endl;
 	vector<float> gaussWidths = identification.getGaussWidths();
 	a.resize(gaussWidths.size());
-	for (int centerId=0; centerId<gaussWidths.size(); centerId++)
+	for (size_t centerId=0; centerId<gaussWidths.size(); centerId++)
 		a[centerId].resize(train_data.getNumFeatures(), gaussWidths[centerId]);
 	b0 = identification.getVectorB();
 }
@@ -30,7 +30,7 @@ float ModelTSK0::getAnswer(vector<float> x)
 	for (int i=0; i<num_rules; i++)
 	{
 		float out = 1.f;
-		for (int j=0; j<x.size(); j++)
+		for (size_t j=0; j<x.size(); j++)
 		{
 			float e = exp(-(x[j] - c[i][j])*(x[j] - c[i][j])/(2*a[i][j]));
 			out *= e*e;
@@ -49,6 +49,6 @@ float ModelTSK0::getAnswer(vector<float> x)
 
 	// calculate answer
 	float result = sum1/sum2;
-	return floor(result + 0.5);
+	return (float) floor(result + 0.5);
     //return result;
 }
